+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,4 +1,8 @@
-﻿### Version 3.10.1
+﻿### Version 3.10.2
+
+- Fixed: Mapping system not correctly handling Queries which return a collection and have more than one parameters.
+
+### Version 3.10.1
 
 - Improvement: Mappings with missing target elements will have a better looking exception.
 - Fixed: Mapping system not correctly handling primary keys when working with inheritance.